Improve branches selections responsive
